Create a simple interface for the steganography functions
This change simplifies the interface for the main() function. Only
hide() and reveal() are exposed. The process of figuring out which
functionality to  perform is handled by these two, which is why the
struct Args is passed as a parameter. Also, the rest of the specific
functions (hide_*() and reveal_*() family) are properly converted into
static.