Use safe function for subtracting unsigned integers
This change implements safe_subtract() which is a function to safely
perform subtraction between two size_t's with regards to underflows. If
an underflow is detected, the function returns false and the program
will display an error message and abort.

Underflows will usually occur when an image is very small, e.g. when it
has one pixel. When hiding a file within such an image, the program will
write 4 length bytes for the size of the file to hide. So in this case,
one (size of image) subtracted by four (length bytes of file), will
trigger the underflow because the result will be in a size_t. See
commit diff for a better picture.